refactor: move top-level components
